Condense FromStream/Extend impls
