--- src/collections/binary_heap/mod.rs
@@ -1,7 +1,0 @@
-//! The Rust priority queue implemented with a binary heap
-
-mod extend;
-mod from_stream;
-
-#[doc(inline)]
-pub use std::collections::BinaryHeap;
