--- src/collections/btree_map/mod.rs
@@ -1,7 +1,0 @@
-//! The Rust B-Tree Map
-
-mod extend;
-mod from_stream;
-
-#[doc(inline)]
-pub use std::collections::BTreeMap;
