--- src/collections/btree_set/mod.rs
@@ -1,7 +1,0 @@
-//! The Rust B-Tree Set
-
-mod extend;
-mod from_stream;
-
-#[doc(inline)]
-pub use std::collections::BTreeSet;
