--- src/collections/linked_list/mod.rs
@@ -1,7 +1,0 @@
-//! The Rust doubly-linked list with owned nodes
-
-mod extend;
-mod from_stream;
-
-#[doc(inline)]
-pub use std::collections::LinkedList;
