--- src/option/mod.rs
@@ -1,9 +1,0 @@
-//! The Rust core optional value type
-//!
-//! This module provides the `Option<T>` type for returning and
-//! propagating optional values.
-
-mod from_stream;
-
-#[doc(inline)]
-pub use std::option::Option;
