--- src/result/mod.rs
@@ -1,9 +1,0 @@
-//! The Rust core error handling type
-//!
-//! This module provides the `Result<T, E>` type for returning and
-//! propagating errors.
-
-mod from_stream;
-
-#[doc(inline)]
-pub use std::result::Result;
