--- src/unit/mod.rs
@@ -1,6 +1,0 @@
-//! The Rust primitive `()` type, sometimes called "unit" or "nil".
-//!
-//! This module provides types and implementations for working
-//! asynchronously with values of type `()`.
-
-mod from_stream;
