--- src/vec/mod.rs
@@ -1,10 +1,0 @@
-//! The Rust core allocation and collections library
-//!
-//! This library provides smart pointers and collections for managing
-//! heap-allocated values.
-
-mod extend;
-mod from_stream;
-
-#[doc(inline)]
-pub use std::vec::Vec;
